refactor: change operatorService.ts to operator-service.ts to match repo naming patterns
